+++ src/moscot/problems/base/_base_problem.py
@@ -17,7 +17,7 @@ from moscot.problems._utils import wrap_solve, wrap_prepare, require_solution
 from moscot.solvers._output import BaseSolverOutput
 from moscot.solvers._base_solver import OTSolver, ProblemKind
 from moscot._constants._constants import ProblemStage
-from moscot.solvers._tagged_array import Tag, TaggedArray
+from moscot.problems._tagged_array import Tag, TaggedArray
 
 __all__ = ["BaseProblem", "OTProblem", "ProblemKind"]
