+++ src/moscot/solvers/__init__.py
@@ -1,3 +1,2 @@
 from moscot.solvers._output import BaseSolverOutput, MatrixSolverOutput
 from moscot.solvers._base_solver import OTSolver, BaseSolver
-from moscot.solvers._tagged_array import Tag, TaggedArray
