+++ tests/backends/ott/test_backend.py
@@ -19,7 +19,7 @@ from moscot._types import Device_t, ArrayLike
 from moscot.backends.ott import GWSolver, SinkhornSolver  # type: ignore[attr-defined]
 from moscot.solvers._output import BaseSolverOutput
 from moscot.solvers._base_solver import O, OTSolver
-from moscot.solvers._tagged_array import Tag
+from moscot.problems._tagged_array import Tag
 
 
 class TestSinkhorn:
